Added calendar and price for rooms

+++ database/migrations/2025_06_23_173925_create_menus_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('menus', function (Blueprint $table) {
             $table->id();
-            $table->integer('price');
+            $table->decimal('price', 10, 2);
             $table->foreignId('user_id')->constrained('users')->onDelete('cascade');
             $table->foreignId('restaurant_id')->nullable()->constrained('restaurants')->onDelete('cascade');
         });
